Changed names of tests from SP to MP, changed function exactE in TestGaussian_convergence

+++ src/Solver/test/mixed_precision/TestGaussian_convergence.cpp
@@ -13,7 +13,7 @@
 using Mesh_t = ippl::UniformCartesian<double, 3>;
 using Centering_t = Mesh_t::DefaultCentering;
 using ScalarField_t = ippl::Field<double, 3, Mesh_t, Centering_t>;
-using VectorField_t = ippl::Field<ippl::Vector<double, 3>, 3, Mesh_t, Centering_t>;
+using VectorField_t = ippl::Field<ippl::Vector<float, 3>, 3, Mesh_t, Centering_t>;
 
 KOKKOS_INLINE_FUNCTION double gaussian(double x, double y, double z, double sigma = 0.05,
                                        double mu = 0.5) {
@@ -32,15 +32,15 @@ KOKKOS_INLINE_FUNCTION double exact_fct(double x, double y, double z, double sig
     return (1 / (4.0 * pi * r)) * std::erf(r / (std::sqrt(2.0) * sigma));
 }
 
-KOKKOS_INLINE_FUNCTION ippl::Vector<double, 3> exact_E(double x, double y, double z,
-                                                       double sigma = 0.05, double mu = 0.5) {
-    double pi     = std::acos(-1.0);
-    double r      = std::sqrt((x - mu) * (x - mu) + (y - mu) * (y - mu) + (z - mu) * (z - mu));
-    double factor = (1.0 / (4.0 * pi * r * r))
+KOKKOS_INLINE_FUNCTION ippl::Vector<float, 3> exact_E(float x, float y, float z,
+                                                       float sigma = 0.05, float mu = 0.5) {
+    float pi     = std::acos(-1.0);
+    float r      = std::sqrt((x - mu) * (x - mu) + (y - mu) * (y - mu) + (z - mu) * (z - mu));
+    float factor = (1.0 / (4.0 * pi * r * r))
                     * ((1.0 / r) * std::erf(r / (std::sqrt(2.0) * sigma))
                        - std::sqrt(2.0 / pi) * (1.0 / sigma) * exp(-r * r / (2 * sigma * sigma)));
 
-    ippl::Vector<double, 3> Efield = {(x - mu), (y - mu), (z - mu)};
+    ippl::Vector<float, 3> Efield = {(x - mu), (y - mu), (z - mu)};
     return factor * Efield;
 }
 
@@ -199,9 +199,9 @@ int main(int argc, char* argv[]) {
                 const int jg = j + ldom[1].first() - nghost;
                 const int kg = k + ldom[2].first() - nghost;
 
-                double x = (ig + 0.5) * hx[0] + origin[0];
-                double y = (jg + 0.5) * hx[1] + origin[1];
-                double z = (kg + 0.5) * hx[2] + origin[2];
+                float x = (ig + 0.5) * float( hx[0] + origin[0] );
+                float y = (jg + 0.5) * float( hx[1] + origin[1] );
+                float z = (kg + 0.5) * float( hx[2] + origin[2] );
 
                 view_exactE(i, j, k)[0] = exact_E(x, y, z)[0];
                 view_exactE(i, j, k)[1] = exact_E(x, y, z)[1];
@@ -217,7 +217,7 @@ int main(int argc, char* argv[]) {
         fftParams.add("comm", ippl::a2av);
         fftParams.add("r2c_direction", 0);
         // define an FFTPoissonSolver object
-        ippl::FFTPoissonSolver<ippl::Vector<double, 3>, double, 3, Mesh_t, Centering_t> FFTsolver(fieldE,
+        ippl::FFTPoissonSolver<ippl::Vector<float, 3>, double, 3, Mesh_t, Centering_t> FFTsolver(fieldE,
                                                                                                   rho,
                                                                                                   fftParams,
                                                                                                   algorithm);
@@ -230,27 +230,27 @@ int main(int argc, char* argv[]) {
         double err = norm(rho) / norm(exact);
 
         // compute relative error norm for the E-field components
-        ippl::Vector<double, 3> errE{0.0, 0.0, 0.0};
+        ippl::Vector<float, 3> errE{0.0, 0.0, 0.0};
         fieldE           = fieldE - exactE;
         auto view_fieldE = fieldE.getView();
 
         for (size_t d = 0; d < 3; ++d) {
-            double temp = 0.0;
+            float temp = 0.0;
             Kokkos::parallel_reduce(
                 "Vector errorNr reduce",
                 Kokkos::MDRangePolicy<Kokkos::Rank<3>>(
                     {nghost, nghost, nghost},
                     {view_fieldE.extent(0) - nghost, view_fieldE.extent(1) - nghost,
                      view_fieldE.extent(2) - nghost}),
-                KOKKOS_LAMBDA(const size_t i, const size_t j, const size_t k, double& valL) {
+                KOKKOS_LAMBDA(const size_t i, const size_t j, const size_t k, float& valL) {
                     double myVal = pow(view_fieldE(i, j, k)[d], 2);
                     valL += myVal;
                 },
-                Kokkos::Sum<double>(temp));
+                Kokkos::Sum<float>(temp));
 
-            double globaltemp = 0.0;
+            float globaltemp = 0.0;
             MPI_Allreduce(&temp, &globaltemp, 1, MPI_DOUBLE, MPI_SUM, Ippl::getComm());
-            double errorNr = std::sqrt(globaltemp);
+            float errorNr = std::sqrt(globaltemp);
 
             temp = 0.0;
             Kokkos::parallel_reduce(
@@ -259,15 +259,15 @@ int main(int argc, char* argv[]) {
                     {nghost, nghost, nghost},
                     {view_exactE.extent(0) - nghost, view_exactE.extent(1) - nghost,
                      view_exactE.extent(2) - nghost}),
-                KOKKOS_LAMBDA(const size_t i, const size_t j, const size_t k, double& valL) {
-                    double myVal = pow(view_exactE(i, j, k)[d], 2);
+                KOKKOS_LAMBDA(const size_t i, const size_t j, const size_t k, float& valL) {
+                    float myVal = pow(view_exactE(i, j, k)[d], 2);
                     valL += myVal;
                 },
-                Kokkos::Sum<double>(temp));
+                Kokkos::Sum<float>(temp));
 
             globaltemp = 0.0;
             MPI_Allreduce(&temp, &globaltemp, 1, MPI_DOUBLE, MPI_SUM, Ippl::getComm());
-            double errorDr = std::sqrt(globaltemp);
+            float errorDr = std::sqrt(globaltemp);
 
             errE[d] = errorNr / errorDr;
         }
